P3 command example
